Remake

+++ src/main/java/Ints.java
@@ -1,5 +1,5 @@
 public interface Ints {
-    Double sum(int arg0, int arg1);
-    Double mult(int arg0, int arg1);
-    Double pow(double a, double b);
+    int sum(int arg0, int arg1);
+    int mult(int arg0, int arg1);
+    int pow(double a, double b);
 }
